Update README.md
This is my first Bi-weekly assessment during the Nsonye Cohort 3. The instructions and dataset was provided by the course tutor, Mr. Gbolahan Adebayo. The skills that was taught and applied to this assessment was Pivot tables, Inserting calculated fields, Conditional Formatting, Filters, Sorting, Grouping, I also checking the dataset for duplicates. My solutions and steps in answering the questions is uploaded and I also publised a the answers and steps in Medium. you can access the article using this link